+++ src/main/scala/org/masood/mafia/repository/SessionRepository.scala
@@ -13,5 +13,5 @@ class SessionRepository(val redisTemplate: RedisTemplate[String, Session]) {
     hashOperations.put("SESSION", session.userId.toString, session)
   }
 
-  def findById(id: Int) = Option(hashOperations.get("SESSION", id.toString))
+  def findById(id: Long) = Option(hashOperations.get("SESSION", id.toString))
 }
